+++ device/api/umd/device/arch/blackhole_implementation.h
@@ -12,7 +12,7 @@
 #include "umd/device/arch/architecture_implementation.h"
 #include "umd/device/types/cluster_descriptor_types.h"
 #include "umd/device/types/tlb.h"
-#include "umd/device/umd_utils.h"
+#include "umd/device/utils/common.h"
 
 namespace tt::umd {
