+++ device/api/umd/device/pcie/pci_device.hpp
@@ -15,11 +15,11 @@
 #include <vector>
 
 #include "fmt/format.h"
-#include "umd/device/semver.hpp"
-#include "umd/device/tt_device/tlb_handle.h"
+#include "umd/device/pcie/tlb_handle.h"
 #include "umd/device/tt_xy_pair.h"
 #include "umd/device/types/arch.h"
 #include "umd/device/types/tlb.h"
+#include "umd/device/utils/semver.hpp"
 
 namespace tt::umd {
 class semver_t;
