+++ device/api/umd/device/pcie/tlb_window.h
@@ -7,7 +7,7 @@
 
 #include <memory>
 
-#include "umd/device/tt_device/tlb_handle.h"
+#include "umd/device/pcie/tlb_handle.h"
 
 namespace tt::umd {
