+++ device/api/umd/device/types/tensix_soft_reset_options.hpp
