+++ device/chip/chip.cpp
@@ -13,8 +13,8 @@
 #include "umd/device/arch/architecture_implementation.h"
 #include "umd/device/arch/wormhole_implementation.h"
 #include "umd/device/driver_atomics.h"
-#include "umd/device/pci_device.hpp"
-#include "umd/device/tt_silicon_driver_common.hpp"
+#include "umd/device/pcie/pci_device.hpp"
+#include "umd/device/types/tensix_soft_reset_options.hpp"
 
 extern bool umd_use_noc1;
