+++ tests/pcie/test_pcie_device.cpp
@@ -12,7 +12,7 @@
 #include <vector>
 
 #include "fmt/xchar.h"
-#include "umd/device/pci_device.hpp"
+#include "umd/device/pcie/pci_device.hpp"
 
 using namespace tt::umd;
